move belief viewer chapter to ch12, close #918
