+++ src/ch12_belief_viewer/_ref/ch12_ref.json
@@ -1,6 +1,6 @@
 {
   "chapter_blurb": "Tools for Visualizing BeliefUnits", 
   "chapter_content": "Tools for Visualizing BeliefUnits", 
-  "chapter_description": "ch22_belief_viewer", 
-  "chapter_number": 22
+  "chapter_description": "ch12_belief_viewer", 
+  "chapter_number": 12
 }
