+++ src/ch12_belief_viewer/example22_beliefs.py
