+++ src/ch12_belief_viewer/static/app.js
