+++ src/ch12_belief_viewer/static/style.css
