+++ src/ch12_belief_viewer/test/_util/ch12_env.py
@@ -4,16 +4,16 @@ from typing import Any, Generator, Literal
 
 
 def get_chapter_temp_dir() -> str:
-    return "src/ch22_belief_viewer"
+    return "src/ch12_belief_viewer"
 
 
-def get_chapter_temp_dir() -> Literal["src\\ch22_belief_viewer\\test\\_util\\temp"]:
-    return "src\\ch22_belief_viewer\\test\\_util\\temp"
+def get_chapter_temp_dir() -> Literal["src\\ch12_belief_viewer\\test\\_util\\temp"]:
+    return "src\\ch12_belief_viewer\\test\\_util\\temp"
 
 
 @pytest_fixture()
 def env_dir_setup_cleanup() -> (
-    Generator[Literal["src\\ch22_belief_viewer\\test\\_util"], Any, None]
+    Generator[Literal["src\\ch12_belief_viewer\\test\\_util"], Any, None]
 ):
     env_dir = get_chapter_temp_dir()
     delete_dir(dir=env_dir)
