+++ src/ch12_belief_viewer/test/_util/example_belief1.json
