+++ src/ch12_belief_viewer/test/test__plan_view_dict.py
@@ -3,8 +3,8 @@ from src.ch08_epoch.reason_str_func import (
     get_fact_state_readable_str,
     get_reason_case_readable_str,
 )
-from src.ch22_belief_viewer.belief_viewer__tool import add_small_dot, get_plan_view_dict
-from src.ch22_belief_viewer.example22_beliefs import (
+from src.ch12_belief_viewer.belief_viewer__tool import add_small_dot, get_plan_view_dict
+from src.ch12_belief_viewer.example22_beliefs import (
     best_run_str,
     best_soccer_str,
     best_sport_str,
@@ -17,7 +17,7 @@ from src.ch22_belief_viewer.example22_beliefs import (
     play_str,
     play_swim_str,
 )
-from src.ref.keywords import Ch22Keywords as wx
+from src.ref.keywords import Ch12Keywords as wx
 
 
 def test_get_plan_view_dict_ReturnsObj_Scenario0_EmptyPlan():
