+++ src/ch12_belief_viewer/test/test__voices_view_dict.py
@@ -1,9 +1,9 @@
 from src.ch07_belief_logic.belief_main import beliefunit_shop
-from src.ch22_belief_viewer.belief_viewer__tool import (
+from src.ch12_belief_viewer.belief_viewer__tool import (
     add_small_dot,
     get_voices_view_dict,
 )
-from src.ref.keywords import Ch22Keywords as wx
+from src.ref.keywords import Ch12Keywords as wx
 
 
 def test_get_voices_view_dict_ReturnsObj_Scenario0_Empty():
