+++ src/ch12_belief_viewer/test/test_belief_viewer_readable.py
@@ -3,13 +3,13 @@ from src.ch08_epoch.reason_str_func import (
     get_fact_state_readable_str,
     get_reason_case_readable_str,
 )
-from src.ch22_belief_viewer.belief_viewer__tool import (
+from src.ch12_belief_viewer.belief_viewer__tool import (
     add_small_dot,
     get_belief_view_dict,
     get_plan_view_dict,
     get_voices_view_dict,
 )
-from src.ref.keywords import Ch22Keywords as wx
+from src.ref.keywords import Ch12Keywords as wx
 
 
 def test_get_belief_view_dict_ReturnsObj_Scenario0_Empty():
