+++ src/ref/keywords.py
@@ -1217,6 +1217,213 @@ class Ch11Keywords(str, Enum):
         return self.value
 
 
+class Ch12Keywords(str, Enum):
+    BeliefName = "BeliefName"
+    CRUD_command = "CRUD_command"
+    DELETE = "DELETE"
+    EpochLabel = "EpochLabel"
+    EpochPoint = "EpochPoint"
+    EventInt = "EventInt"
+    FaceName = "FaceName"
+    FirstLabel = "FirstLabel"
+    FundGrain = "FundGrain"
+    FundNum = "FundNum"
+    GrainNum = "GrainNum"
+    GroupTitle = "GroupTitle"
+    HealerName = "HealerName"
+    INSERT = "INSERT"
+    KnotTerm = "KnotTerm"
+    LabelTerm = "LabelTerm"
+    MomentLabel = "MomentLabel"
+    MoneyGrain = "MoneyGrain"
+    MoneyNum = "MoneyNum"
+    NameTerm = "NameTerm"
+    NexusLabel = "NexusLabel"
+    PoolNum = "PoolNum"
+    RespectGrain = "RespectGrain"
+    RespectNum = "RespectNum"
+    RopeTerm = "RopeTerm"
+    TitleTerm = "TitleTerm"
+    UPDATE = "UPDATE"
+    VoiceName = "VoiceName"
+    active = "active"
+    active_hx = "active_hx"
+    addin = "addin"
+    all_voice_cred = "all_voice_cred"
+    all_voice_debt = "all_voice_debt"
+    amount = "amount"
+    ancestors = "ancestors"
+    atom = "atom"
+    atom_hx = "atom_hx"
+    awardee_title = "awardee_title"
+    awardheirs = "awardheirs"
+    awardlines = "awardlines"
+    awardunits = "awardunits"
+    begin = "begin"
+    belief_groupunit = "belief_groupunit"
+    belief_name = "belief_name"
+    belief_plan_awardunit = "belief_plan_awardunit"
+    belief_plan_factunit = "belief_plan_factunit"
+    belief_plan_healerunit = "belief_plan_healerunit"
+    belief_plan_partyunit = "belief_plan_partyunit"
+    belief_plan_reason_caseunit = "belief_plan_reason_caseunit"
+    belief_plan_reasonunit = "belief_plan_reasonunit"
+    belief_planunit = "belief_planunit"
+    belief_voice_membership = "belief_voice_membership"
+    belief_voiceunit = "belief_voiceunit"
+    beliefadjust = "beliefadjust"
+    beliefevent_facts = "beliefevent_facts"
+    beliefunit = "beliefunit"
+    boss_facts = "boss_facts"
+    bud_belief_name = "bud_belief_name"
+    bud_time = "bud_time"
+    bud_voice_nets = "bud_voice_nets"
+    c100 = "c100"
+    c400_clean = "c400_clean"
+    c400_leap = "c400_leap"
+    c400_number = "c400_number"
+    cases = "cases"
+    celldepth = "celldepth"
+    class_type = "class_type"
+    close = "close"
+    column_order = "column_order"
+    credor_pool = "credor_pool"
+    credor_respect = "credor_respect"
+    creg = "creg"
+    crud = "crud"
+    cumulative_day = "cumulative_day"
+    day = "day"
+    days = "days"
+    debtor_pool = "debtor_pool"
+    debtor_respect = "debtor_respect"
+    denom = "denom"
+    descendant_pledge_count = "descendant_pledge_count"
+    dimen = "dimen"
+    dimens = "dimens"
+    epoch = "epoch"
+    epoch_label = "epoch_label"
+    event_int = "event_int"
+    face_name = "face_name"
+    fact_context = "fact_context"
+    fact_lower = "fact_lower"
+    fact_state = "fact_state"
+    fact_upper = "fact_upper"
+    factheirs = "factheirs"
+    factunits = "factunits"
+    five = "five"
+    found_facts = "found_facts"
+    fund_agenda_give = "fund_agenda_give"
+    fund_agenda_ratio_give = "fund_agenda_ratio_give"
+    fund_agenda_ratio_take = "fund_agenda_ratio_take"
+    fund_agenda_take = "fund_agenda_take"
+    fund_cease = "fund_cease"
+    fund_give = "fund_give"
+    fund_grain = "fund_grain"
+    fund_onset = "fund_onset"
+    fund_pool = "fund_pool"
+    fund_ratio = "fund_ratio"
+    fund_take = "fund_take"
+    give_force = "give_force"
+    gogo_calc = "gogo_calc"
+    gogo_want = "gogo_want"
+    group_cred_lumen = "group_cred_lumen"
+    group_debt_lumen = "group_debt_lumen"
+    group_title = "group_title"
+    groupunits = "groupunits"
+    gut = "gut"
+    healer_name = "healer_name"
+    healerunit = "healerunit"
+    healerunit_ratio = "healerunit_ratio"
+    hour = "hour"
+    hours_config = "hours_config"
+    inallocable_voice_debt_lumen = "inallocable_voice_debt_lumen"
+    irrational_voice_debt_lumen = "irrational_voice_debt_lumen"
+    is_expanded = "is_expanded"
+    jkeys = "jkeys"
+    job = "job"
+    jvalues = "jvalues"
+    keep_rope = "keep_rope"
+    keeps_buildable = "keeps_buildable"
+    keeps_justified = "keeps_justified"
+    kids = "kids"
+    knot = "knot"
+    laborheir = "laborheir"
+    laborunit = "laborunit"
+    last_pack_id = "last_pack_id"
+    magnitude = "magnitude"
+    mandate = "mandate"
+    max_tree_traverse = "max_tree_traverse"
+    memberships = "memberships"
+    moment_label = "moment_label"
+    moment_mstr_dir = "moment_mstr_dir"
+    money_grain = "money_grain"
+    monthday_index = "monthday_index"
+    months_config = "months_config"
+    morph = "morph"
+    nesting_order = "nesting_order"
+    nexus = "nexus"
+    normal_specs = "normal_specs"
+    normal_table_name = "normal_table_name"
+    numor = "numor"
+    offi_time = "offi_time"
+    offi_time_max = "offi_time_max"
+    offtrack_fund = "offtrack_fund"
+    offtrack_kids_star_set = "offtrack_kids_star_set"
+    parent_rope = "parent_rope"
+    parent_solo = "parent_solo"
+    party_title = "party_title"
+    plan_fund_total = "plan_fund_total"
+    plan_label = "plan_label"
+    plan_rope = "plan_rope"
+    planroot = "planroot"
+    pledge = "pledge"
+    problem_bool = "problem_bool"
+    quota = "quota"
+    range_evaluated = "range_evaluated"
+    rational = "rational"
+    readable = "readable"
+    reason_active_requisite = "reason_active_requisite"
+    reason_context = "reason_context"
+    reason_contexts = "reason_contexts"
+    reason_divisor = "reason_divisor"
+    reason_lower = "reason_lower"
+    reason_state = "reason_state"
+    reason_upper = "reason_upper"
+    reasonheirs = "reasonheirs"
+    reasonunits = "reasonunits"
+    respect_grain = "respect_grain"
+    solo = "solo"
+    sqlite_datatype = "sqlite_datatype"
+    star = "star"
+    status = "status"
+    stop_calc = "stop_calc"
+    stop_want = "stop_want"
+    sum_healerunit_plans_fund_total = "sum_healerunit_plans_fund_total"
+    take_force = "take_force"
+    tally = "tally"
+    task = "task"
+    time = "time"
+    tran_time = "tran_time"
+    tree_level = "tree_level"
+    tree_traverse_count = "tree_traverse_count"
+    uid = "uid"
+    voice_cred_lumen = "voice_cred_lumen"
+    voice_debt_lumen = "voice_debt_lumen"
+    voice_name = "voice_name"
+    voice_pool = "voice_pool"
+    voices = "voices"
+    week = "week"
+    weekdays_config = "weekdays_config"
+    weeks = "weeks"
+    year = "year"
+    yr1_jan1_offset = "yr1_jan1_offset"
+    yr4_clean = "yr4_clean"
+    yr4_leap = "yr4_leap"
+
+    def __str__(self):
+        return self.value
+
+
 class Ch13Keywords(str, Enum):
     BeliefName = "BeliefName"
     CRUD_command = "CRUD_command"
@@ -3378,313 +3585,6 @@ class Ch20Keywords(str, Enum):
 
 
 class Ch21Keywords(str, Enum):
-    BeliefName = "BeliefName"
-    CRUD_command = "CRUD_command"
-    DELETE = "DELETE"
-    EpochLabel = "EpochLabel"
-    EpochPoint = "EpochPoint"
-    EventInt = "EventInt"
-    FaceName = "FaceName"
-    FirstLabel = "FirstLabel"
-    FundGrain = "FundGrain"
-    FundNum = "FundNum"
-    GrainNum = "GrainNum"
-    GroupTitle = "GroupTitle"
-    HealerName = "HealerName"
-    INSERT = "INSERT"
-    KnotTerm = "KnotTerm"
-    LabelTerm = "LabelTerm"
-    LobbyID = "LobbyID"
-    MomentLabel = "MomentLabel"
-    MoneyGrain = "MoneyGrain"
-    MoneyNum = "MoneyNum"
-    NameTerm = "NameTerm"
-    NexusLabel = "NexusLabel"
-    PoolNum = "PoolNum"
-    RespectGrain = "RespectGrain"
-    RespectNum = "RespectNum"
-    RopeTerm = "RopeTerm"
-    TitleTerm = "TitleTerm"
-    UPDATE = "UPDATE"
-    VoiceName = "VoiceName"
-    WorldName = "WorldName"
-    _credor_count = "_credor_count"
-    _cycle_chargeees_curr = "_cycle_chargeees_curr"
-    _cycle_chargeees_prev = "_cycle_chargeees_prev"
-    _cycle_count = "_cycle_count"
-    _debtor_count = "_debtor_count"
-    _grants = "_grants"
-    _rivergrades = "_rivergrades"
-    _rivergrants = "_rivergrants"
-    _tax_got_curr = "_tax_got_curr"
-    _tax_got_prev = "_tax_got_prev"
-    _tax_yields = "_tax_yields"
-    active = "active"
-    active_hx = "active_hx"
-    addin = "addin"
-    all_tranbook = "all_tranbook"
-    all_voice_cred = "all_voice_cred"
-    all_voice_debt = "all_voice_debt"
-    allowed_crud = "allowed_crud"
-    amount = "amount"
-    ancestors = "ancestors"
-    atom = "atom"
-    atom_hx = "atom_hx"
-    attributes = "attributes"
-    awardee_title = "awardee_title"
-    awardheirs = "awardheirs"
-    awardlines = "awardlines"
-    awardunits = "awardunits"
-    begin = "begin"
-    belief_groupunit = "belief_groupunit"
-    belief_name = "belief_name"
-    belief_net_amount = "belief_net_amount"
-    belief_plan_awardunit = "belief_plan_awardunit"
-    belief_plan_factunit = "belief_plan_factunit"
-    belief_plan_healerunit = "belief_plan_healerunit"
-    belief_plan_partyunit = "belief_plan_partyunit"
-    belief_plan_reason_caseunit = "belief_plan_reason_caseunit"
-    belief_plan_reasonunit = "belief_plan_reasonunit"
-    belief_planunit = "belief_planunit"
-    belief_voice_membership = "belief_voice_membership"
-    belief_voiceunit = "belief_voiceunit"
-    beliefadjust = "beliefadjust"
-    beliefbudhistorys = "beliefbudhistorys"
-    beliefevent_facts = "beliefevent_facts"
-    beliefunit = "beliefunit"
-    boss_facts = "boss_facts"
-    brick_agg = "brick_agg"
-    brick_raw = "brick_raw"
-    brick_valid = "brick_valid"
-    bud_belief_name = "bud_belief_name"
-    bud_time = "bud_time"
-    bud_voice_nets = "bud_voice_nets"
-    build_order = "build_order"
-    c100 = "c100"
-    c400_clean = "c400_clean"
-    c400_leap = "c400_leap"
-    c400_number = "c400_number"
-    cases = "cases"
-    celldepth = "celldepth"
-    class_type = "class_type"
-    close = "close"
-    column_order = "column_order"
-    credor_count = "credor_count"
-    credor_pool = "credor_pool"
-    credor_rank_num = "credor_rank_num"
-    credor_rank_percent = "credor_rank_percent"
-    credor_respect = "credor_respect"
-    creg = "creg"
-    crud = "crud"
-    cumulative_day = "cumulative_day"
-    cumulative_minute = "cumulative_minute"
-    cycle_max = "cycle_max"
-    day = "day"
-    days = "days"
-    debtor_count = "debtor_count"
-    debtor_pool = "debtor_pool"
-    debtor_rank_num = "debtor_rank_num"
-    debtor_rank_percent = "debtor_rank_percent"
-    debtor_respect = "debtor_respect"
-    default_kpi_bundle = "default_kpi_bundle"
-    delete_insert = "delete_insert"
-    delete_insert_update = "delete_insert_update"
-    delete_update = "delete_update"
-    denom = "denom"
-    descendant_pledge_count = "descendant_pledge_count"
-    dimen = "dimen"
-    dimens = "dimens"
-    epoch = "epoch"
-    epoch_label = "epoch_label"
-    error_message = "error_message"
-    event_int = "event_int"
-    events_brick_agg = "events_brick_agg"
-    events_brick_valid = "events_brick_valid"
-    face_name = "face_name"
-    fact_context = "fact_context"
-    fact_lower = "fact_lower"
-    fact_state = "fact_state"
-    fact_upper = "fact_upper"
-    factheirs = "factheirs"
-    factunits = "factunits"
-    five = "five"
-    found_facts = "found_facts"
-    fund_agenda_give = "fund_agenda_give"
-    fund_agenda_ratio_give = "fund_agenda_ratio_give"
-    fund_agenda_ratio_take = "fund_agenda_ratio_take"
-    fund_agenda_take = "fund_agenda_take"
-    fund_cease = "fund_cease"
-    fund_give = "fund_give"
-    fund_grain = "fund_grain"
-    fund_onset = "fund_onset"
-    fund_pool = "fund_pool"
-    fund_ratio = "fund_ratio"
-    fund_take = "fund_take"
-    give_force = "give_force"
-    gogo_calc = "gogo_calc"
-    gogo_want = "gogo_want"
-    grant_amount = "grant_amount"
-    group_cred_lumen = "group_cred_lumen"
-    group_debt_lumen = "group_debt_lumen"
-    group_title = "group_title"
-    groupunits = "groupunits"
-    gut = "gut"
-    healer_name = "healer_name"
-    healerunit = "healerunit"
-    healerunit_ratio = "healerunit_ratio"
-    heard_agg = "heard_agg"
-    heard_raw = "heard_raw"
-    hour = "hour"
-    hour_label = "hour_label"
-    hours_config = "hours_config"
-    idea_category = "idea_category"
-    idea_number = "idea_number"
-    inallocable_voice_debt_lumen = "inallocable_voice_debt_lumen"
-    insert_multiple = "insert_multiple"
-    insert_one_time = "insert_one_time"
-    insert_update = "insert_update"
-    inx_knot = "inx_knot"
-    inx_label = "inx_label"
-    inx_name = "inx_name"
-    inx_rope = "inx_rope"
-    inx_title = "inx_title"
-    irrational_voice_debt_lumen = "irrational_voice_debt_lumen"
-    is_expanded = "is_expanded"
-    jkeys = "jkeys"
-    job = "job"
-    job_listen_rotations = "job_listen_rotations"
-    jvalues = "jvalues"
-    keep_credorledgers = "keep_credorledgers"
-    keep_point_magnitude = "keep_point_magnitude"
-    keep_rope = "keep_rope"
-    keeps_buildable = "keeps_buildable"
-    keeps_justified = "keeps_justified"
-    kids = "kids"
-    knot = "knot"
-    laborheir = "laborheir"
-    laborunit = "laborunit"
-    last_pack_id = "last_pack_id"
-    lobby_id = "lobby_id"
-    lobby_mstr_dir = "lobby_mstr_dir"
-    lobbys = "lobbys"
-    magnitude = "magnitude"
-    mandate = "mandate"
-    max_tree_traverse = "max_tree_traverse"
-    memberships = "memberships"
-    moment_budunit = "moment_budunit"
-    moment_epoch_hour = "moment_epoch_hour"
-    moment_epoch_month = "moment_epoch_month"
-    moment_epoch_weekday = "moment_epoch_weekday"
-    moment_event_time_agg = "moment_event_time_agg"
-    moment_kpi001_voice_nets = "moment_kpi001_voice_nets"
-    moment_kpi002_belief_pledges = "moment_kpi002_belief_pledges"
-    moment_label = "moment_label"
-    moment_mstr_dir = "moment_mstr_dir"
-    moment_ote1_agg = "moment_ote1_agg"
-    moment_paybook = "moment_paybook"
-    moment_timeoffi = "moment_timeoffi"
-    moment_voice_nets = "moment_voice_nets"
-    momentunit = "momentunit"
-    money_grain = "money_grain"
-    month_label = "month_label"
-    monthday_index = "monthday_index"
-    months_config = "months_config"
-    morph = "morph"
-    nesting_order = "nesting_order"
-    nexus = "nexus"
-    normal_specs = "normal_specs"
-    normal_table_name = "normal_table_name"
-    numor = "numor"
-    offi_time = "offi_time"
-    offi_time_max = "offi_time_max"
-    offi_times = "offi_times"
-    offtrack_fund = "offtrack_fund"
-    offtrack_kids_star_set = "offtrack_kids_star_set"
-    otx2inx = "otx2inx"
-    otx_key = "otx_key"
-    otx_knot = "otx_knot"
-    otx_label = "otx_label"
-    otx_name = "otx_name"
-    otx_rope = "otx_rope"
-    otx_title = "otx_title"
-    parent_rope = "parent_rope"
-    parent_solo = "parent_solo"
-    party_title = "party_title"
-    paybook = "paybook"
-    plan_fund_total = "plan_fund_total"
-    plan_label = "plan_label"
-    plan_rope = "plan_rope"
-    planroot = "planroot"
-    pledge = "pledge"
-    problem_bool = "problem_bool"
-    quota = "quota"
-    range_evaluated = "range_evaluated"
-    rational = "rational"
-    readable = "readable"
-    reason_active_requisite = "reason_active_requisite"
-    reason_context = "reason_context"
-    reason_contexts = "reason_contexts"
-    reason_divisor = "reason_divisor"
-    reason_lower = "reason_lower"
-    reason_state = "reason_state"
-    reason_upper = "reason_upper"
-    reasonheirs = "reasonheirs"
-    reasonunits = "reasonunits"
-    respect_grain = "respect_grain"
-    rewards_count = "rewards_count"
-    rewards_magnitude = "rewards_magnitude"
-    riverbooks = "riverbooks"
-    solo = "solo"
-    sound_agg = "sound_agg"
-    sound_raw = "sound_raw"
-    sqlite_datatype = "sqlite_datatype"
-    star = "star"
-    status = "status"
-    stop_calc = "stop_calc"
-    stop_want = "stop_want"
-    sum_healerunit_plans_fund_total = "sum_healerunit_plans_fund_total"
-    take_force = "take_force"
-    tally = "tally"
-    task = "task"
-    tax_bill_amount = "tax_bill_amount"
-    tax_dues = "tax_dues"
-    tax_paid_amount = "tax_paid_amount"
-    tax_paid_bool = "tax_paid_bool"
-    tax_paid_rank_num = "tax_paid_rank_num"
-    tax_paid_rank_percent = "tax_paid_rank_percent"
-    time = "time"
-    tran_time = "tran_time"
-    translate_core = "translate_core"
-    translate_label = "translate_label"
-    translate_name = "translate_name"
-    translate_rope = "translate_rope"
-    translate_title = "translate_title"
-    translateunit = "translateunit"
-    tree_level = "tree_level"
-    tree_traverse_count = "tree_traverse_count"
-    uid = "uid"
-    unknown_str = "unknown_str"
-    voice_cred_lumen = "voice_cred_lumen"
-    voice_debt_lumen = "voice_debt_lumen"
-    voice_name = "voice_name"
-    voice_pool = "voice_pool"
-    voices = "voices"
-    week = "week"
-    weekday_label = "weekday_label"
-    weekday_order = "weekday_order"
-    weekdays_config = "weekdays_config"
-    weeks = "weeks"
-    world_name = "world_name"
-    year = "year"
-    yr1_jan1_offset = "yr1_jan1_offset"
-    yr4_clean = "yr4_clean"
-    yr4_leap = "yr4_leap"
-
-    def __str__(self):
-        return self.value
-
-
-class Ch22Keywords(str, Enum):
     BeliefName = "BeliefName"
     CRUD_command = "CRUD_command"
     DELETE = "DELETE"
